Increase attachment_parser_service test coverage
Additionally fixed what I think was a bug, where if some of the attachments failed to parse (missing keys) we log an error but the return result of this logging was being added to the collection (an integer), which I presume is not what we want here.

I've changed the code slightly so the rescue returns `nil`, and using `filter_map` to take care of these nil values automatically.

I've added a test for this edge case scenario too.