Refactor to make methods private
And propagate the `request_id ` to the Job.

Removed the capture_message with the slug, because the params when the exception is captured in sentry will contain the slug so is redundant.

Disable `wrap_parameters` as we did also in the filestore service, as this duplicates the payloads, is not neccessary.