Prepare to remove PDF from confirmation email
There is already a flag to indicate if we want the PDF with the answers generated or not / attached to the emails.

We will be using this flag to disable the PDF from the confirmation email.

Runner will need to send false in this flag in order for the answers PDF to not be attached to the confirmation emails.