Support JSON objects in answers (addresses)
We need to support payloads with answers being hashes, and properly translate it for the CSV and the PDF.

This so far only applies to address components. Up until now we were sending in the payload a string concatenating the parts of the address, but we need a structured hash with each of the individual fields so API adapters receive it intact.