Add submission email body heading
Includes the name of the service, the ID and the email counter.

A new `variant` JSON attribute will be needed to propagate from the runner to know exactly which kind of email it is, submission or confirmation, as doing it in any other way is becoming too cumbersome and fragile.